cmdline: move helper types into separate file
